melhorias no processamento de pagamentos

+++ src/RinhaBackend.Api/Application/Response/PaymentProcessorHealthResponse.cs
@@ -1,3 +1,3 @@
 ﻿namespace RinhaBackend.Api.Application.Response;
 
-public record PaymentProcessorHealthResponse(bool Failing, decimal MinResponseTime);
+public record PaymentProcessorHealthResponse(bool Failing, int MinResponseTime);
